2 FA

+++ app/Http/Controllers/TwoFactorAuthController.php
@@ -7,18 +7,26 @@ use App\Models\User;
 use App\Rules\TwoFactorCodeRule;
 use Illuminate\Contracts\Support\Responsable;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\RateLimiter;
 use Symfony\Component\HttpFoundation\Response;
 
 class TwoFactorAuthController extends Controller
 {
     public function setup(): Responsable
     {
-        $secret = Authenticator::generateSecret();
+        if (session()->has('google2fa_secret')) {
+            $secret = session()->get('google2fa_secret');
+        } else {
+            session()->put('google2fa_secret', $secret = Authenticator::generateSecret());
+        }
+
         $qr = Authenticator::getQRCodeUrl(config('app.name'), auth()->user()->email, $secret);
 
-        session()->flash('google2fa_secret', $secret);
+        $debug_code = config('google2fa.debug') && config('app.debug')
+            ? Authenticator::getCurrentOTP($secret)
+            : null;
 
-        return inertia('2fa/setup', compact('secret', 'qr'));
+        return inertia('2fa/setup', compact('secret', 'qr', 'debug_code'));
     }
 
     public function store(Request $request)
@@ -43,12 +51,25 @@ class TwoFactorAuthController extends Controller
         $recovery_codes = Authenticator::generateRecoveryCodes();
 
         $user->recovery_codes()
-            ->insert(array_map(fn($code) => ['code' => $code], $recovery_codes));
+            ->insert(array_map(fn($code) => ['code' => $code, 'user_id' => $user->id], $recovery_codes));
 
         session()->forget('google2fa_secret');
 
+        Authenticator::forceSession();
+
         return redirect()->route('profile')
             ->with('recovery_codes', $recovery_codes);
+    }
+
+    public function verifyPage(): Responsable
+    {
+        return inertia('2fa/verify',
+            [
+                'debug_code' => config('google2fa.debug') && config('app.debug')
+                    ? Authenticator::getCurrentOTP(auth()->user()->google2fa_secret)
+                    : null,
+            ]
+        );
     }
 
     public function verify(Request $request): Response
@@ -70,12 +91,31 @@ class TwoFactorAuthController extends Controller
         /** @var User $user */
         $user = $request->user();
 
+        if (RateLimiter::tooManyAttempts("disable-2fa:$user->id", 5)) {
+            return back()->with(['alert' => [
+                'title' => 'Too many attempts',
+                'content' => 'Please try again later.',
+            ]]);
+        }
+
+        RateLimiter::hit("disable-2fa:$user->id", 300);
+
+        $request->validate([
+            'password' => ['required', 'current_password'],
+        ]);
+
+        RateLimiter::clear("disable-2fa:$user->id");
+
         $user->update([
             'google2fa_secret' => null,
         ]);
 
         $user->recovery_codes()->delete();
 
-        return redirect()->route('profile.index');
+        return redirect()->route('profile')
+            ->with('alert', [
+                'title' => 'Success',
+                'content' => 'Two-factor authentication has been disabled.',
+            ]);
     }
 }
